refactor: remove scoped tojos class
