+++ eo-maven-plugin/src/main/java/org/eolang/maven/RegisterMojo.java
@@ -110,7 +110,7 @@ public final class RegisterMojo extends SafeMojo {
     @Override
     public void exec() throws IOException {
         final Pattern pattern = Pattern.compile("^[a-zA-Z0-9\\-]+\\.eo$");
-        final int before = this.scopedTojos().toForeignTojos().size();
+        final int before = this.scopedTojos().size();
         if (before > 0) {
             Logger.info(this, "There are %d EO sources registered already", before);
         }
